kb: revision of renewing Harvester kubeconfig
Revision of KB article for renewing expired Harvester cloud
credentials/kubeconfig.
This revision updates the procedure to make it slightly easier to yield
the correct kubeconfig file by instructing users to use the existing
mechanisms in the UI.
Additionally, this saves on complicated processing in the patch
scriptlet.
The screenshots have been updated to give a more complete view of the
web UI and show the context menu where the user would find the button to
generate a new kubeconfig file.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>